@@ -1,27 +1,27 @@
 <template>
-  <div class="flex message d-flex message-bot messsge-loading" aria-hidden="true">
+  <v-row d-flex class="message message-bot messsge-loading" aria-hidden="true">
     <!-- contains message and response card -->
-    <v-layout column ma-2 class="message-layout">
+    <v-col ma-2 class="message-layout">
 
       <!-- contains message bubble and date -->
-      <v-flex d-flex class="message-bubble-date-container">
-        <v-layout column class="message-bubble-column">
+      <v-row d-flex class="message-bubble-date-container">
+        <v-col class="message-bubble-column">
 
           <!-- contains message bubble and avatar -->
-          <v-flex d-flex class="message-bubble-avatar-container">
-            <v-layout row class="message-bubble-row">
+          <v-col d-flex class="message-bubble-avatar-container">
+            <v-row class="message-bubble-row">
               <div
                 class="message-bubble"
                 aria-hidden="true"
               >
               {{$store.state.config.lex.allowStreamingResponses? $store.state.streaming.wsMessagesString : progress }}
               </div>
-            </v-layout>
-          </v-flex>
-        </v-layout>
-      </v-flex>
-    </v-layout>
-  </div>
+            </v-row>
+          </v-col>
+        </v-col>
+      </v-row>
+    </v-col>
+  </v-row>
 </template>
 
 <script>
@@ -61,7 +61,7 @@ export default {
       }
     }, 500);
   },
-  destroyed() {
+  unmounted() {
     clearInterval(this.interval);
   },
 };
