@@ -1,100 +1,90 @@
 <template>
-  <div app fixed>
-    <v-layout
-      row
-      justify-space-between
-      ma-0
-      class="input-container"
-    >
-      <v-toolbar
-        color="white"
-        v-bind:dense="this.$store.state.isRunningEmbedded"
+  <v-toolbar elevation="3" color="white" :dense="this.$store.state.isRunningEmbedded" class="toolbar-content">
+    <!--
+      using v-show instead of v-if to make recorder-status transition work
+    -->
+      <!--
+        using v-show instead of v-if to make recorder-status transition work
+      -->
+      <v-text-field
+        :label="textInputPlaceholder"
+        v-show="shouldShowTextInput"
+        :disabled="isLexProcessing"
+        v-model="textInput"
+        @keyup.enter.stop="postTextMessage"
+        @focus="onTextFieldFocus"
+        @blur="onTextFieldBlur"
+        @update:model-value="onKeyUp"
+        ref="textInput"
+        id="text-input"
+        name="text-input"
+        single-line
+        hide-details
+        density="compact"
+        variant="underlined"
+        class="toolbar-text"
       >
-        <!--
-          using v-show instead of v-if to make recorder-status transition work
-        -->
-        <v-text-field
-          v-bind:label="textInputPlaceholder"
-          v-show="shouldShowTextInput"
-          v-bind:disabled="isLexProcessing"
-          v-model="textInput"
-          v-on:keyup.enter.stop="postTextMessage"
-          v-on:focus="onTextFieldFocus"
-          v-on:blur="onTextFieldBlur"
-          @input="onKeyUp"
-          ref="textInput"
-          id="text-input"
-          name="text-input"
-          single-line
-          hide-details
-        ></v-text-field>
+    </v-text-field>
 
-        <recorder-status
-          v-show="!shouldShowTextInput"
-        ></recorder-status>
+      <recorder-status
+        v-show="!shouldShowTextInput"
+      ></recorder-status>
 
-        <!-- separate tooltip as a workaround to support mobile touch events -->
-        <!-- tooltip should be before btn to avoid right margin issue in mobile -->
-        <v-tooltip
-          activator=".input-button"
-          content-class="tooltip-custom"
-          v-model="shouldShowTooltip"
-          ref="tooltip"
-          left
-        >
-          <span id="input-button-tooltip">{{inputButtonTooltip}}</span>
-        </v-tooltip>
-        <v-btn
-          v-if="shouldShowSendButton"
-          v-on:click="postTextMessage"
-          v-on="tooltipEventHandlers"
-          v-bind:disabled="isLexProcessing || isSendButtonDisabled"
-          ref="send"
-          class="icon-color input-button"
-          icon
-          aria-label="Send Message"
-        >
-          <v-icon medium>send</v-icon>
-        </v-btn>
-        <v-btn
-          v-if="!shouldShowSendButton && !isModeLiveChat"
-          v-on:click="onMicClick"
-          v-on="tooltipEventHandlers"
-          v-bind:disabled="isMicButtonDisabled"
-          ref="mic"
-          class="icon-color input-button"
-          icon
-        >
-          <v-icon medium>{{micButtonIcon}}</v-icon>
-        </v-btn>
-        <v-btn
-          v-if="shouldShowUpload"
-          v-on:click="onPickFile"
-          v-bind:disabled="isLexProcessing"
-          ref="upload"
-          class="icon-color input-button"
-          icon
-        >
-          <v-icon medium>attach_file</v-icon>
-          <input
-            type="file"
-            style="display: none"
-            ref="fileInput"
-            @change="onFilePicked">
-        </v-btn>
-        <v-btn
-          v-if="shouldShowAttachmentClear"
-          v-on:click="onRemoveAttachments"
-          v-bind:disabled="isLexProcessing"
-          ref="removeAttachments"
-          class="icon-color input-button"
-          icon
-        >
-          <v-icon medium>clear</v-icon>
-        </v-btn>
-      </v-toolbar>
-    </v-layout>
-  </div>
+            <!-- separate tooltip as a workaround to support mobile touch events -->
+    <!-- tooltip should be before btn to avoid right margin issue in mobile -->
+    <v-btn
+      v-if="shouldShowSendButton"
+      @click="postTextMessage"
+      :disabled="isLexProcessing || isSendButtonDisabled"
+      ref="send"
+      class="icon-color input-button"
+      aria-label="Send Message"
+    >
+      <v-tooltip activator="parent" location="start">
+        <span id="input-button-tooltip">{{ inputButtonTooltip }}</span>
+      </v-tooltip>
+      <v-icon size="x-large">send</v-icon>
+    </v-btn>
+    <v-btn
+      v-if="!shouldShowSendButton && !isModeLiveChat"
+      @click="onMicClick"
+      v-on="tooltipEventHandlers"
+      :disabled="isMicButtonDisabled"
+      ref="mic"
+      class="icon-color input-button"
+      icon
+    >
+      <v-tooltip activator="parent" v-model="shouldShowTooltip" location="start">
+        <span id="input-button-tooltip">{{ inputButtonTooltip }}</span>
+      </v-tooltip>
+      <v-icon size="x-large">{{ micButtonIcon }}</v-icon>
+    </v-btn>
+    <v-btn
+      v-if="shouldShowUpload"
+      v-on:click="onPickFile"
+      v-bind:disabled="isLexProcessing"
+      ref="upload"
+      class="icon-color input-button"
+      icon
+    >
+      <v-icon size="x-large">attach_file</v-icon>
+      <input
+        type="file"
+        style="display: none"
+        ref="fileInput"
+        @change="onFilePicked">
+    </v-btn>
+    <v-btn
+      v-if="shouldShowAttachmentClear"
+      v-on:click="onRemoveAttachments"
+      v-bind:disabled="isLexProcessing"
+      ref="removeAttachments"
+      class="icon-color input-button"
+      icon
+    >
+      <v-icon size="x-large">clear</v-icon>
+    </v-btn>
+  </v-toolbar>
 </template>
 
 <script>
@@ -235,7 +225,7 @@ export default {
       // focus() needs to be wrapped in setTimeout for IE11
       setTimeout(() => {
         if (this.$refs && this.$refs.textInput && this.shouldShowTextInput) {
-          this.$refs.textInput.$refs.input.focus();
+          this.$refs.textInput.focus();
         }
       }, 10);
     },
@@ -365,4 +355,14 @@ export default {
   right: 0;
   right: env(safe-area-inset-right);
 }
+
+.toolbar-content {
+  padding-left: 16px;
+  font-size: 16px !important;
+}
+
+.v-input {
+  margin-bottom: 10px;
+}
+
 </style>
