@@ -74,6 +74,18 @@ function chainWebpackWorker(config, destDir = '', srcDir = 'src/lib') {
     .use('babel-loader')
     .loader('babel-loader')
     .end();
+
+  // custom components
+  config.module
+    .rule('vue')
+    .use('vue-loader')
+    .tap(options => {
+      options.compilerOptions = {
+        ...options.compilerOptions,
+        isCustomElement: tag => tag.startsWith('v-datetime-picker')
+      }
+      return options
+    });
 }
 
 function chainWebpackCommon(config, destDir) {
@@ -97,6 +109,15 @@ function chainWebpackCommon(config, destDir) {
       return args;
     })
     .end();
+
+  config.plugin('define').tap((definitions) => {
+    Object.assign(definitions[0], {
+      __VUE_OPTIONS_API__: 'true',
+      __VUE_PROD_DEVTOOLS__: 'false',
+      __VUE_PROD_HYDRATION_MISMATCH_DETAILS__: 'false'
+    })
+    return definitions
+  })
 
   config.plugin('NodePolyfillPlugin').use(new NodePolyfillPlugin());
 }
@@ -128,12 +149,14 @@ function chainWebpackLib(
     // XXX TODO need to add dependencies below to the lex-web-ui-loader
     // 'jsonwebtoken',
     // 'marked',
-    'vue',
-    'vuex',
+    { 'vue': 'Vue' },
+    {'vuex': 'Vuex'},
     'vue-router',
-    'vuetify',
+    {'vuetify': 'Vuetify'},
     /^aws-sdk\/.+$/,
   ]);
+
+  config.externalsType = 'window';
 
   config.optimization.splitChunks({
     cacheGroups: {
@@ -188,7 +211,17 @@ function chainWebpackLib(
     }]);
 }
 
-function chainWebpackApp(config, destDir = '') {
+function chainWebpackApp(
+  config,
+  entryName = 'lex-web-ui',
+  entryFileName = './src/lex-web-ui.js',
+  destDir = ''
+) {
+
+  config
+    .entry(entryName)
+    .add(entryFileName);
+
   config.output.filename(
     (buildType.isProd) ? '[name].min.js' : '[name].js',
   );
