@@ -60,6 +60,9 @@ module.exports = (env) => {
           test: /\.js$/,
           exclude: /[\\/]node_modules[\\/]/,
           loader: 'babel-loader',
+          options: {
+            presets: ['@babel/preset-env']
+          }
         },
         {
           test: /\.css$/,
@@ -123,12 +126,14 @@ module.exports = (env) => {
         template: path.join(basePath, 'src/website/index.html'),
         // script is included in template
         inject: false,
+        scriptLoading: 'blocking',
       }),
       new HtmlWebpackPlugin({
         filename: 'parent.html',
         template: path.join(basePath, 'src/website/parent.html'),
         // script is included in template
         inject: false,
+        scriptLoading: 'blocking',
       }),
       isProd && new webpack.BannerPlugin({
         banner: `/*!
@@ -147,10 +152,10 @@ module.exports = (env) => {
         {
           patterns: [
             // copy parent page
-            {
-              from: path.join(basePath, 'src/website/parent.html'),
-              to: distDir,
-            },
+            //{
+            //  from: path.join(basePath, 'src/website/parent.html'),
+            //  to: distDir,
+            //},
             // copy custom css
             {
               from: path.join(basePath, 'src/website/custom-chatbot-style.css'),
@@ -171,5 +176,10 @@ module.exports = (env) => {
         }
       ),
     ].filter(Boolean),
+    performance: {
+      hints: false,
+      maxEntrypointSize: 512000,
+      maxAssetSize: 512000
+  },
   };
 };
